pyproject.toml
[tool.poetry]
name = "support-project-radar"
version = "0.1.0"
description = "Automated grant and funding opportunity radar."
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pydantic = "^1.10.0"
requests = "^2.28.0"
feedparser = "^6.0.10"
sqlite-utils = "^3.31.0"
slack-sdk = "^3.19.0"
pyyaml = "^6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"